[hv] hv: Send one page worth of kmsg dump over Hyper-V during panic
Message-id: <20190506164615.3165-6-mgamal@redhat.com>
Patchwork-id: 253836
O-Subject: [RHEL8.1 KERNEL PATCH 05/32] Drivers: HV: Send one page worth of kmsg dump over Hyper-V during panic
Bugzilla: 1691383
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Cathy Avery <cavery@redhat.com>

In the VM mode on Hyper-V, currently, when the kernel panics, an error
code and few register values are populated in an MSR and the Hypervisor
notified. This information is collected on the host. The amount of
information currently collected is found to be limited and not very
actionable. To gather more actionable data, such as stack trace, the
proposal is to write one page worth of kmsg data on an allocated page
and the Hypervisor notified of the page address through the MSR.

- Sysctl option to control the behavior, with ON by default.

Cc: K. Y. Srinivasan <kys@microsoft.com>
Cc: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: Sunil Muthuswamy <sunilmut@microsoft.com>
Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 81b18bce48af3e222ca9eebb3b931f59141b7b49)

Signed-off-by: Mohammed Gamal <mgamal@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>