[tools] tools/power turbostat: Restore ability to execute in topology-order
Message-id: <20190506192527.29524.79762.sendpatchset@stevebest>
Patchwork-id: 253879
O-Subject: [PATCH RHEL8.1 BZ1705959 04/14] tools/power turbostat: Restore ability to execute in topology-order
Bugzilla: 1705959
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1705959

Description:
------------
turbostat executes on CPUs in "topology order".
This is an optimization for measuring profoundly idle systems --
as the closest hardware is woken next...

Fix a typo that was added with the sub-die-node support,
that broke topology ordering on multi-node systems.

Signed-off-by: Len Brown <len.brown@intel.com>

Upstream Status:
----------------
commit df2f677dee3ce4afda7ee561dd4f321c40320afd

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=df2f677dee3ce4afda7ee561dd4f321c40320afd

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>