[fs] cifs: remove coverity warning in calc_lanman_hash
Message-id: <20190509235548.8854-128-lsahlber@redhat.com>
Patchwork-id: 254430
O-Subject: [RHEL8.1 fs 127/276] cifs: remove coverity warning in calc_lanman_hash
Bugzilla: 1708033
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1708033
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=21579898

password_with_pad is a fixed size buffer of 16 bytes, it contains a
password string, to be padded with \0 if shorter than 16 bytes
but is just truncated if longer.
It is not, and we do not depend on it to be, nul terminated.

As such, do not use strncpy() to populate this buffer since
the str* prefix suggests that this is a string, which it is not,
and it also confuses coverity causing a false warning.

Detected by CoverityScan CID#113743 ("Buffer not null terminated")

Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
(cherry picked from commit 52baa51d30573273341b78a3e3ed0d9a845d3e55)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>