[rpmspec] Update the kernel rpm package description to mention backports/bug fixes
Message-id: <20190514145914.21051-1-herton@redhat.com>
Patchwork-id: 255151
O-Subject: [RHEL8 PATCH v3] [rpmspec] Update the kernel rpm package description to mention backports/bug fixes
Bugzilla: 1701064
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1701064
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=21577076
Tested: Verified package summary/description of built package
Upstream Status: RHEL only

There is a misconception on the press and probably with some users that RHEL ships
with an old/outdated kernel version. That's not true, because we backport a lot of
features and bug fixes from newer kernels. Our package shows the version of the
upstream kernel we based initially, and we can't change that because we are doing
selected fixes/backports, and changing/updating this version while not updating the
entire kernel would be lying about it too.

The changelog already shows what we backport, but the entries don't point to the
upstream kernel version the changes are coming from.

Brendan Conoboy made a suggestion to improve this situation: update the summary and
description of the kernel meta package. This change tries to enhance the summary and
description to make it clear that the kernel is not a rebuild close to a pristine
upstream release, but instead contains several backports and changes that may bring
it in line with latest upstream releases in many areas.

v2: add "... while maintaining a well tested and stable core." to the description
as suggested by Luis Claudio R. Goncalves; minor tweaks.
v3: fix kABI wording as suggested by Jiri Benc

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>