[kvm] KVM: PPC: Propagate errors to the guest when failed instead of ignoring
Message-id: <20190514173034.32627-1-vkuznets@redhat.com>
Patchwork-id: 255263
O-Subject: [RHEL8.1 virt PATCH v3 033/549] KVM: PPC: Propagate errors to the guest when failed instead of ignoring
Bugzilla: 1692804
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Alexey Kardashevskiy <aik@ozlabs.ru>

At the moment if the PUT_TCE{_INDIRECT} handlers fail to update
the hardware tables, we print a warning once, clear the entry and
continue. This is so as at the time the assumption was that if
a VFIO device is hotplugged into the guest, and the userspace replays
virtual DMA mappings (i.e. TCEs) to the hardware tables and if this fails,
then there is nothing useful we can do about it.

However the assumption is not valid as these handlers are not called for
TCE replay (VFIO ioctl interface is used for that) and these handlers
are for new TCEs.

This returns an error to the guest if there is a request which cannot be
processed. By now the only possible failure must be H_TOO_HARD.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(cherry picked from commit 2691f0ff3d96b471a69c7ef29f3340b91056845d)

RHEL-only: out of order c10c21efa4bc backport requires kvmppc_clear_tce()
call sites adjustments.

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>